testing with mock data and implemented test calc for total

+++ src/main/groovy/com/cagnosolutions/starter/app/item/Item.groovy
@@ -19,7 +19,7 @@ class Item {
 	Long id
 	@OneToOne
 	Material material
-	Integer count, total
+	Float count, total
 
 	void calcTotal() {
 		this.total = count * material.price
